pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[project]
name = "retix"
version = "1.2.0"
description = "RETIX - The Optic Nerve for Autonomous Agents"
authors = [{ name = "Vision Team" }]
readme = "README.md"
license = "MIT"
requires-python = ">=3.10,<4.0"
dependencies = [
    "click>=8.1.7",
    "rich-click>=1.7.0",
    "rich>=13.7.0",
    "pillow>=10.1.0",
    "numpy>=1.24.0,<2.0",
    "mlx>=0.25.0",
    "mlx-vlm>=0.1.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0",
    "psutil>=5.9.6",
    "numexpr>=2.10.2",
    "bottleneck>=1.4.2",
]

[project.scripts]
retix = "retix.main:cli"

[tool.poetry]
name = "retix"
version = "1.2.0"
description = "RETIX - The Optic Nerve for Autonomous Agents"
authors = ["Vision Team <dev@retix.local>"]
readme = "README.md"
license = "MIT"

packages = [
    { include = "retix" }
]

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.7"
rich-click = "^1.7.0"
rich = "^13.7.0"
pillow = "^10.1.0"
numpy = ">=1.24.0,<2.0"
mlx = ">=0.25.0"
mlx-vlm = ">=0.1.0"
torch = ">=2.0.0"
torchvision = ">=0.15.0"
pydantic = "^2.5.0"
pyyaml = "^6.0"
psutil = "^5.9.6"
numexpr = ">=2.10.2"
bottleneck = ">=1.4.2"

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "^23.11.0"
ruff = "^0.1.8"
mypy = "^1.7.0"
types-pyyaml = "^6.0.12"

[tool.poetry.scripts]
retix = "retix.main:cli"

[tool.black]
line-length = 100
target-version = ["py39"]
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py39"
select = ["E", "F", "W", "I"]
ignore = ["E501"]  # Line too long - handled by black

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = false
ignore_missing_imports = true
